+++ curve25519-dalek/CHANGELOG.md
@@ -5,6 +5,11 @@ major series.
 
 ## 5.x series
 
+## 5.0.0-pre.3
+
+* Add Lizard bytes-to-point injection for Ristretto. Gated under `lizard` feature.
+* Upgrade `rand_core` to v0.10.0-rc-2
+
 ## 5.0.0-pre.1
 
 * Rename `Scalar::batch_invert` -> `Scalar::invert_batch` for consistency. Also make it no-alloc.