+++ ed25519-dalek/CHANGELOG.md
@@ -8,6 +8,10 @@ Entries are listed in reverse chronological order per undeprecated major series.
 
 # 3.x series
 
+## 3.0.0-pre.3
+
+* Upgrade `rand_core` to v0.10.0-rc-2
+
 ## 3.0.0-pre.0
 
 * Update edition to 2024