+++ x25519-dalek/CHANGELOG.md
@@ -4,6 +4,14 @@ Entries are listed in reverse chronological order.
 
 # 3.x Series
 
+## Unreleased
+
+* Removed `alloc` feature flag, which was doing nothing
+
+## 3.0.0-pre.3
+
+* Upgrade `rand_core` to v0.10.0-rc-2
+
 ## 3.0.0-pre.0
 
 * Update edition to 2024